refactor: split code
split the dom implementation into their own modules
